+++ BinanceExchange.API/Client/ClientConfiguration.cs
@@ -1,3 +1,5 @@
+using NLog;
+
 namespace BinanceExchange.API.Client
 {
     public class ClientConfiguration
@@ -5,5 +7,6 @@ namespace BinanceExchange.API.Client
         public string ApiKey { get; set; }
         public string SecretKey { get; set; }
         public bool EnableRateLimiting { get; set; }
+        public ILogger Logger { get; set; }
     }
 }