+++ BinanceExchange.API/RequestClient.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 using BinanceExchange.API.Enums;
+using NLog;
 
 namespace BinanceExchange.API
 {
@@ -25,6 +26,7 @@ namespace BinanceExchange.API
         private const string APIHeader = "X-MBX-APIKEY";
         private static readonly Stopwatch Stopwatch;
         private static int _concurrentRequests = 0;
+        private static ILogger _logger;
 
         static RequestClient()
         {
@@ -36,6 +38,7 @@ namespace BinanceExchange.API
             HttpClient.DefaultRequestHeaders.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
             _rateSemaphore = new SemaphoreSlim(_limit, _limit);
             Stopwatch = new Stopwatch();
+            _logger = LogManager.GetCurrentClassLogger();
         }
 
 
@@ -47,6 +50,7 @@ namespace BinanceExchange.API
         {
             _limit = limit;
             _rateSemaphore = new SemaphoreSlim(limit, limit);
+            _logger.Debug($"Request Limit Adjusted to: {limit}");
         }
 
         /// <summary>
@@ -55,7 +59,9 @@ namespace BinanceExchange.API
         /// <param name="enabled"></param>
         public static void SetRateLimiting(bool enabled)
         {
+            var set = enabled ? "enabled" : "disabled";
             RateLimitingEnabled = enabled;
+            _logger.Debug($"Rate Limiting has been {set}");
         }
 
         /// <summary>
@@ -65,6 +71,7 @@ namespace BinanceExchange.API
         public static void SetSecondsLimit(int limit)
         {
             SecondsLimit = limit;
+            _logger.Debug($"Rate Limiting seconds limit has been set to {limit}");
         }
 
         /// <summary>
@@ -87,6 +94,7 @@ namespace BinanceExchange.API
         /// <returns></returns>
         public static async Task<HttpResponseMessage> GetRequest(Uri endpoint)
         {
+            _logger.Debug($"Creating a GET Request to {endpoint.AbsoluteUri}");
             return await CreateRequest(endpoint);
         }
 
@@ -101,6 +109,7 @@ namespace BinanceExchange.API
         /// <returns></returns>
         public static async Task<HttpResponseMessage> SignedGetRequest(Uri endpoint, string apiKey, string secretKey, string signatureRawData, long receiveWindow = 5000)
         {
+            _logger.Debug($"Creating a SIGNED GET Request to {endpoint.AbsoluteUri}");
             var uri = CreateValidUri(endpoint, secretKey, signatureRawData, receiveWindow);
             return await CreateRequest(uri, HttpVerb.GET);
         }
@@ -112,6 +121,7 @@ namespace BinanceExchange.API
         /// <returns></returns>
         public static async Task<HttpResponseMessage> PostRequest(Uri endpoint)
         {
+            _logger.Debug($"Creating a POST Request to {endpoint.AbsoluteUri}");
             return await CreateRequest(endpoint, HttpVerb.POST);
         }
 
@@ -122,6 +132,7 @@ namespace BinanceExchange.API
         /// <returns></returns>
         public static async Task<HttpResponseMessage> DeleteRequest(Uri endpoint)
         {
+            _logger.Debug($"Creating a DELETE Request to {endpoint.AbsoluteUri}");
             return await CreateRequest(endpoint, HttpVerb.DELETE);
         }
 
@@ -132,6 +143,7 @@ namespace BinanceExchange.API
         /// <returns></returns>
         public static async Task<HttpResponseMessage> PutRequest(Uri endpoint)
         {
+            _logger.Debug($"Creating a PUT Request to {endpoint.AbsoluteUri}");
             return await CreateRequest(endpoint, HttpVerb.PUT);
         }
 
@@ -146,6 +158,7 @@ namespace BinanceExchange.API
         /// <returns></returns>
         public static async Task<HttpResponseMessage> SignedPostRequest(Uri endpoint, string apiKey, string secretKey, string signatureRawData, long receiveWindow = 5000)
         {
+            _logger.Debug($"Creating a SIGNED POST Request to {endpoint.AbsoluteUri}");
             var uri = CreateValidUri(endpoint, secretKey, signatureRawData, receiveWindow);
             return await CreateRequest(uri, HttpVerb.POST);
         }
@@ -161,6 +174,7 @@ namespace BinanceExchange.API
         /// <returns></returns>
         public static async Task<HttpResponseMessage> SignedDeleteRequest(Uri endpoint, string apiKey, string secretKey, string signatureRawData, long receiveWindow = 5000)
         {
+            _logger.Debug($"Creating a SIGNED DELETE Request to {endpoint.AbsoluteUri}");
             var uri = CreateValidUri(endpoint, secretKey, signatureRawData, receiveWindow);
             return await CreateRequest(uri, HttpVerb.DELETE);
         }
@@ -256,5 +270,9 @@ namespace BinanceExchange.API
             return await task;
         }
 
+        public static void SetLogger(ILogger logger)
+        {
+            _logger = logger;
+        }
     }
 }